Building slots get added as prosperity grows, building slots now automatically belong to player

+++ src/main/java/hedgehogs/strategyGame/gameLogic/land/prosperity/Prosperity.java
@@ -2,9 +2,13 @@ package hedgehogs.strategyGame.gameLogic.land.prosperity;
 
 public interface Prosperity {
 
-    public int getCurrentValue();
+    public double getCurrentValue();
 
     public void increaseByValue(int value);
 
     public void decreaseByValue(int amount);
+
+    // public void setReachGoal(Double targetGoal);
+
+    // public void addNewListener(ProsperityListener newListener);
 }
